iio: altera_adxcvr: Support more than 4 lanes
The altera_adxcvr driver currently only supports up to 4 lanes. The number
of lanes of the hardware is read from a configuration register and if it
exceeds 4 the driver will exhibit undefined behavior.

Increase the number of supported lanes to 32, that should be enough for a
while.

Also add a check to prevent undefined behavior if the hardware has more
transceivers than what the driver supports.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>
Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>