iio: jesd204: axi_jesd204_{rx,tx}: Store subclass version in the correct field
Commit a6a208363c2c ("iio: jesd204: axi_jesd204_[rx|tx]: Add support for
JESD204 Subclass 0") added support for subclass 0. But it stores the
selected subclass in the jesd_version field instead of the subclass_version
field of the jesd204_{rx,tx}_config struct.

On RX this currently doesn't have any effect and on TX it means that the
wrong values for the jesd version and subclass version are send during the
ILAS.

Correct this by storing the subclass version in the right field.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>