iio: adc: adar1000: Fix multiple device communication
There are 2 related issues that are highlighted in these scenarios:
1. a) When a read is performed on an ADAR1000 where the HWADDR used matches
the HWADDR of the ADAR1000, the device responds correctly on the designated
output line.
b) When a read is performed on an ADAR1000 where the HWADDR does not match
the HWADDR of the ADAR1000, but the CSB line is shared, the device will
output all zeros on the designated output line. This is fixed by setting
all non-addressed ADAR1000s to the opposite output style (3-wire vs.
4-wire).

2. a) When an ADAR1000 with HWADDR 0b00 is written to register 0x0000, the
resulting behaviour acts as though a "broadcast all: write was made.
Therefore, any device on the CSB line will now have register 0x0000
overwritten. This is fixed by making the write and then going back and
resetting the value of register 0x0000 to devices with HWADDR 0b01, 0b10,
0b11.
b) This issue does not affect other HWADDRs, and is only present when
writing to register 0x0000.

Fixes: dc22d612ff024 ("iio: adc: adar1000: Initial driver version")
Signed-off-by: Mircea Caprioru <mircea.caprioru@analog.com>
(cherry picked from commit e7baffd24c9b64ca2aed1382c8bb87517d3a3f30)