ci: stop auto-sync-ing altera_4.14 & adi-iio branches
It seems that backporting to an older kernel version is a bit more
problematic sometimes, as master is on 4.19 and intel/altera on 4.14.

We should focus on moving support for intel/altera boards in master and not
carry extra branches around.
At some point we may also switch rpi-4.19.y to master, but first we should
start with altera branches.

If we need things in this branch, we should backport it manually.

The adi-iio branch is too far ahead of master and takes some effort to keep
it working as things change in the kernel (like GPIO APIs).

In the near future, kernel 5.4.0 will be the next version that will be
used. At that point version 4.14 will be more difficult to backport to.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>
(cherry picked from commit 0382f725151494220eb951ad2bb1f13bb8efac4d)