jesd204: backport devm_clk_get_optional()
This patch should cause a failure on newer kernels when using this
framework. But until then we can use in the JESD204 contexts.

It's a neat helper that allows us to be used in the context of the JESD204
framework until the kernel will support it and we can remove it from here.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>