iio: jesd204: Support for optional (slaved) GT REFCLK
This patch adds support for an optional CPLL/QPLL REFCLK from a
difference source, which rate and state must be in sync with the main
conv clk (GT REFCLK). This feature is used on the ADRV9009-ZU11EG SoM
in combination with the external FMCOMMS8 FMC card, where each HMC7044
clock chip provides it's own GT REFCLK connected to different QUADs in
an identical configuration.

Signed-off-by: Michael Hennerich <michael.hennerich@analog.com>