iio: axi_jesd204_tx: use jesd204_tx_config -> jesd204_link
The jesd204_link struct contains all of the properties supported by
jesd204_rx_config, so just use it and initialize it in the non-framework
use-case.

This also means allocating the lane IDs in the probe function, if the
driver is not operating with the JESD204 framework.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>
(cherry picked from commit 922ffb35325746ae5b4b58238abeec3993df1567)