iio: buffer-dma: use the bytes_used field when checking for available bytes
When checking that there is data available for reading, we should probably
use the bytes_used field, since that is the actual data that was
transferred, which could be zero or whatever was transferred.

For partial transfers this is important, as we would always return the
size of the block and not the amount that was actually read.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>